pyproject.toml
[project]
name = "smart-parking-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "lucasantonete@gmail.com", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "setuptools>=72.1.0",
    "djangorestframework>=3.15.2",
    "django-safedelete>=1.4.0",
    "django-simple-history>=3.7.0",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.27.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
dev = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
pre_migrate = "python src/manage.py makemigrations"
migrate = "python src/manage.py migrate"
shell = "python src/manage.py shell"
shellp= "python src/manage.py shell_plus"
test = "python src/manage.py test"
populate = "python src/manage.py populate"
loaddata = "python src/manage.py loaddata core.json"
dumpdata = "python src/manage.py dumpdata --indent 2"